requirements.txt
scipion-pyworkflow
scipion-ed
